refactor(migration): streamline codebase for STREAMLINING_PLAN compliance
- Removed Black from pre-commit hooks to eliminate conflicting formatting guidance, establishing Ruff as the canonical linter and formatter.
- Introduced new .ruff.toml configuration for linting and formatting, aligning with the project's goals for code quality.
- Updated CONTRIBUTING.md to emphasize adherence to the STREAMLINING_PLAN and clarified the development workflow.
- Added pyproject.toml for dependency management, marking it as the authoritative source over requirements.txt.
- Implemented compatibility shims in services.api to facilitate migration to the new src/ml_platform_api structure, ensuring backward compatibility during the transition.
- Enhanced CI workflows to reflect the updated structure and improve security scanning processes.
- Added tests to validate the new structure and ensure functionality remains intact.
