+++ apps/api/__init__.py
@@ -1,0 +1,1 @@
+"""Apps API package - Canonical FastAPI entrypoint wrapper."""
