+++ apps/api/main.py
@@ -1,0 +1,3 @@
+"""Canonical API entrypoint; currently delegates to services.api.main during migration."""
+
+from services.api.main import app as app
