+++ mypy.ini
@@ -2,26 +2,58 @@
 # https://mypy.readthedocs.io/en/stable/config_file.html
 
 [mypy]
-python_version = 3.13
+# Align with pyproject.toml [project.requires-python]
+python_version = 3.11
 warn_return_any = True
 warn_unused_configs = True
 check_untyped_defs = True
+# Use explicit package bases to avoid module name conflicts
+# This ensures packages are resolved correctly when src/ is in mypy_path
+explicit_package_bases = True
 
-# Exclude paths that shouldn't be type-checked
-exclude = ^(\.venv/|\.git/|\.pytest_cache/|\.external_modules/|\.terraform/|\.claude/|infra/|platform/|tests/)
+# Search paths:
+# - src: search root for ml_platform_api.* imports (resolves src/ml_platform_api/ as ml_platform_api)
+# - . (current directory): root for all other imports (services.api.*, apps.api.*, tests.*)
+# Note: services, tests, and apps are NOT in mypy_path individually to ensure proper module resolution:
+#   - services/api/ resolves as services.api (not api)
+#   - apps/api/ resolves as apps.api (not api)
+#   - tests/ resolves as tests.* (when imported)
+mypy_path = src:.
 
-# Per-module configuration for libraries lacking type stubs
-# Note: We use targeted ignores instead of blanket --ignore-missing-imports
-# to maintain type checking for the majority of dependencies that ARE typed
+# Explicit files/directories to type-check.
+# Keep focused to avoid surprising failures in infra/policy tooling.
+files =
+    services/api
+    tests
+    src
+    apps
+
+# Exclude non-source / tooling directories from discovery.
+exclude = ^(\.venv/|\.git/|\.pytest_cache/|\.mypy_cache/|\.ruff_cache/|\.external_modules/|\.terraform/|\.claude/)$
+
+# Strict typing for canonical ml_platform_api modules
+[mypy-ml_platform_api.*]
+disallow_untyped_defs = True
+disallow_incomplete_defs = True
+check_untyped_defs = True
+no_implicit_optional = True
+
+# Strict typing for canonical FastAPI entrypoint wrapper
+[mypy-apps.api.*]
+disallow_untyped_defs = True
+check_untyped_defs = True
+
+# Targeted per-module configuration for libraries lacking type stubs.
+# Prefer explicit ignores over blanket --ignore-missing-imports.
 
 [mypy-sklearn.*]
-# scikit-learn has partial native types (added py.typed in v1.3)
-# but scikit-learn-stubs (v0.0.3) is experimental with known bugs
-# Re-evaluate when sklearn 2.x ships with comprehensive type annotations
+# scikit-learn has partial native types; stubs are unstable.
+# Re-evaluate when upstream typing is complete.
 ignore_missing_imports = True
 
 [mypy-skops.*]
 ignore_missing_imports = True
 
 [mypy-pytest.*]
+# Allow pytest internal imports without forcing local stubs.
 ignore_missing_imports = True
