+++ pytest.ini
@@ -1,20 +1,26 @@
 [pytest]
 # Test discovery
+# Current layout: tests/ is the sole root for test collection.
+# Future apps/src tests can be added under tests/ without changing this config.
 testpaths = tests
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
 
 # Coverage configuration
+# Focused on the FastAPI service and canonical ml_platform_api modules.
 addopts =
     --cov=services.api
+    --cov=ml_platform_api
     --cov-report=xml
     --cov-report=term-missing
     -v
 
 [coverage:run]
 # Source code to measure
-source = services/api
+source =
+    services/api
+    src/ml_platform_api
 
 # Paths to exclude from coverage
 omit =
