+++ src/ml_platform_api/__init__.py
@@ -1,0 +1,5 @@
+"""ML Platform API - Canonical implementation package.
+
+This package contains the authoritative implementations for the ML platform.
+Compatibility shims in services.api re-export from here.
+"""
